feature: add new functional for login page
Add new functional for login-page, now you can show or hidden password text. | change alert about error. | refactoring code for loginForm
