+++ src/features/auth/model/types.ts
@@ -10,3 +10,19 @@ export type UserDto = {
   userForResponse: User;
   message: string;
 };
+
+export type LoginFormData = {
+  email: {
+    text: string;
+    error: boolean;
+    color: "primary" | "error";
+    errorMessage: string;
+  };
+  password: {
+    text: string;
+    error: boolean;
+    color: "primary" | "error";
+    errorMessage: string;
+    isVisibleText: boolean;
+  };
+};
